Delete orphan libnss_db.so
For whatever reason, Ubuntu decided to move libnss_db.so out from the libc6 package into its own external package called libnss-db.

However, that package is marked as optional and, despite supposedly being part of glibc, the libc6 package doesn't mention it in its "Depends" section, so it is not automatically pulled along with libc6 when installing through apt. The libc6 package also has a broken symlink inside /usr pointing to that file.

This mess-up is only present on Ubuntu 12.04 i386.

This addresses parts of
https://github.com/AmanoTeam/obggcc/issues/5#issuecomment-3417838867.